Fabric refactor (#368)
* Removed deprecated m2

* Refactor of fabric components

* Temporary pink headers
